chore(contracts): hoist verbatim-duplicate Database<T> generic (GAP-136)
Phase 1 of GAP-136. The 25-line generic Database<T> declaration was
verbatim-duplicated in packages/contracts/src/database/bridge.ts:24-48
and database/type-bridge.ts:23-47 (verified byte-identical). Hoisted
to a new shared module database/types.ts; both files now import + re-
export from there.

Changes:
- NEW: packages/contracts/src/database/types.ts (canonical Database<T>)
- bridge.ts: 33-line block replaced with 2-line import + 1-line re-export
- type-bridge.ts: same pattern
- Net diff -54 LOC of duplication eliminated; single source of truth

Public API unchanged: 'import { Database } from
@revealui/contracts/database/bridge' (or /type-bridge) still works
because both modules re-export the type. No external consumer
breakage.

Phase 2 (renaming the unrelated concrete 'interface Database' in
packages/contracts/src/generated/database.ts to disambiguate from
the generic) is intentionally DEFERRED. That file is auto-generated
('do not hand-edit' header); a rename would require generator-
template change OR post-gen rename step. Tracked separately if/when
prioritized.

Audit: contracts-type-redeclarations.md §2.4. Re-verified byte-
identical declarations on 2026-04-28.

Local verification: pnpm --filter @revealui/contracts typecheck +
test passes (28 test files, 767 tests, all green).

Closes GAP-136 Phase 1 (Phase 2 deferred per gap notes).
